requirements.txt
requests==2.31.0
anthropic==0.39.0
gspread==6.0.0
google-auth==2.25.0
python-dotenv==1.0.0
